+++ strawberryfields/gbs/__init__.py
@@ -11,3 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+"""GBS applications layer"""
+
+import strawberryfields.gbs.clique
+import strawberryfields.gbs.sample
+import strawberryfields.gbs.similarity
+import strawberryfields.gbs.subgraph